+++ max_div/solver/_score.py
@@ -51,7 +51,7 @@ class Score:
     div_tie_breakers: tuple[float, ...]  # diversity tie-breakers - used in case of ties in all higher-prio metrics
 
     # --- helpers ---------------------
-    def as_tuple(self, soft: float = 0.0) -> tuple[float, ...]:
+    def as_tuple(self, soft: float = 0.0, ignore_infeasible_diversity: bool = False) -> tuple[float, ...]:
         """
         Return score as tuple, in order of descending priority, such that tuple-comparison yields correct results.
 
@@ -59,17 +59,23 @@ class Score:
                       0.0 = hard constraints (i.e. constraints score is absolute higher prio than diversity)
                      >0.0 = soft constraints (i.e. diversity score is partly 'mixed into' constraints score)
                                     constraints_soft = constraints^(1-soft) * diversity^soft
+        :param ignore_infeasible_diversity: If `True`, diversity is set to 0.0 if constraints are not fully satisfied.
         """
-        if soft == 0.0:
-            constraint_score = self.constraints  # 100% hard constraints (no influence from diversity)
-        elif soft == 1.0:
-            constraint_score = self.diversity  # 100% soft constraints (ignoring constraint score)
-        elif self.constraints == 0.0 or self.diversity == 0.0:
-            constraint_score = 0.0  # shortcut to avoid zero-division issues & unnecessary **soft computation
+        if ignore_infeasible_diversity and self.constraints < 1.0:
+            # set scores of diversity & tie-breakers to 0.0 in case of infeasible solution
+            # (also, don't perform 'soft constraint' computation, since that also takes into account diversity)
+            return self.size, self.constraints, 0.0, *[0.0 for tb in self.div_tie_breakers]
         else:
-            constraint_score = self.constraints * ((self.diversity / self.constraints) ** soft)
+            if soft == 0.0:
+                constraint_score = self.constraints  # 100% hard constraints (no influence from diversity)
+            elif soft == 1.0:
+                constraint_score = self.diversity  # 100% soft constraints (ignoring constraint score)
+            elif self.constraints == 0.0 or self.diversity == 0.0:
+                constraint_score = 0.0  # shortcut to avoid zero-division issues & unnecessary **soft computation
+            else:
+                constraint_score = self.constraints * ((self.diversity / self.constraints) ** soft)
 
-        return self.size, constraint_score, self.diversity, *self.div_tie_breakers
+            return self.size, constraint_score, self.diversity, *self.div_tie_breakers
 
     # --- math overloads --------------
     def __lt__(self, other: Any) -> bool: